# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: EventHub API
  version: "1.0.0"
  description: "Documentation \"EventHub\" API v1.0"
servers:
  - url: http://localhost:8080
...
paths:
...
  /categories/v1/api:
    post:
      tags:
        - category
      summary: Создать категорию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateDto'
      responses:
        '201':
          description: Категория успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Категория с таким именем уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - category
      summary: Получить список категорий
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Список категорий
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCategoryDtoResponse'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /categories/v1/api/{id}:
    get:
      tags:
        - category
      summary: Получить категорию по id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Категория найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Категория не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - category
      summary: Обновить категорию
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateDto'
      responses:
        '200':
          description: Категория обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Категория не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Категория с таким именем уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - category
      summary: Удалить категорию
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Категория удалена
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Категория не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /categories/v1/api/projects:
    post:
      tags:
        - project
      summary: Создать проект
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateDto'
      responses:
        '201':
          description: Проект успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - project
      summary: Получить список проектов
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedProjectDtoResponse'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /categories/v1/api/projects/{id}:
    get:
      tags:
        - project
      summary: Получить проект по id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Проект найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - project
      summary: Обновить проект
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateDto'
      responses:
        '200':
          description: Проект обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - project
      summary: Удалить проект
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Проект удалён
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CategoryCreateDto:
      type: object
      required: [ name ]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Название категории
          example: "Технологии"
        description:
          type: string
          maxLength: 255
          description: Описание категории
          example: "Категория для технологических событий."
    CategoryUpdateDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: Новое название категории
          example: "Образование"
        description:
          type: string
          maxLength: 255
          description: Новое описание категории
          example: "Категория для образовательных мероприятий."
...
    ProjectCreateDto:
      type: object
      required: [ name, ownerId, categoryId ]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Название проекта
          example: "EventHub"
        description:
          type: string
          maxLength: 255
          description: Описание проекта
          example: "Платформа для организации событий."
        categoryId:
          type: string
          format: uuid
          description: Идентификатор категории
          example: "b1a7e8c2-3d4f-4e2a-9c1a-2b3c4d5e6f7a"
        ownerId:
          type: string
          format: uuid
          description: Идентификатор владельца
          example: "a1234567-89ab-cdef-0123-456789abcdef"
    ProjectUpdateDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: Новое название проекта
          example: "EventHub Pro"
        description:
          type: string
          maxLength: 255
          description: Новое описание проекта
          example: "Расширенная платформа для мероприятий."
        categoryId:
          type: string
          format: uuid
          description: Новый идентификатор категории
          example: "b1a7e8c2-3d4f-4e2a-9c1a-2b3c4d5e6f7a"
        ownerId:
          type: string
          format: uuid
          description: Новый идентификатор владельца
          example: "a1234567-89ab-cdef-0123-456789abcdef"

    # ----- Responses -----
...
    UserShortDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор пользователя
          example: "a1234567-89ab-cdef-0123-456789abcdef"
        username:
          type: string
          description: Имя пользователя
          example: "ivan.petrov"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "ivan.petrov@example.com"
    CategoryShortDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор категории
          example: "b1a7e8c2-3d4f-4e2a-9c1a-2b3c4d5e6f7a"
        name:
          type: string
          description: Название категории
          example: "Технологии"
...
    CategoryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор категории
          example: "b1a7e8c2-3d4f-4e2a-9c1a-2b3c4d5e6f7a"
        name:
          type: string
          description: Название категории
          example: "Технологии"
        description:
          type: string
          description: Описание категории
          example: "Категория для технологических событий."
        createdAt:
          type: string
          format: date-time
          description: Дата создания категории
          example: "2025-12-01T10:00:00+03:00"
        updatedAt:
          type: string
          format: date-time
          description: Дата обновления категории
          example: "2025-12-02T12:00:00+03:00"
    ProjectDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор проекта
          example: "e1f2a3b4-5678-1234-abcd-9876543210ab"
        name:
          type: string
          description: Название проекта
          example: "EventHub"
        description:
          type: string
          description: Описание проекта
          example: "Платформа для организации событий."
        category:
          $ref: '#/components/schemas/CategoryShortDto'
        owner:
          $ref: '#/components/schemas/UserShortDto'
    ErrorResponse:
      type: object
      properties:
        path:
          type: string
          description: Путь запроса, вызвавшего ошибку
          example: "/events"
        httpMethod:
          type: string
          description: HTTP метод запроса
          example: "POST"
        statusCode:
          type: integer
          description: Код ошибки
          example: 400
        errorReason:
          type: string
          description: Краткое описание причины ошибки
          example: "Bad Request"
        message:
          type: string
          description: Описание ошибки
          example: "Некорректные данные запроса"
...
    PagedCategoryDtoResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    PagedProjectDtoResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProjectDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...